backend/requirements.txt
django
djangorestframework
django-cors-headers
pendulum
pillow
psycopg2
pandas
python-dotenv
numpy